Auth Api

+++ src/pages/api/[[...user]].tsx
@@ -16,27 +16,40 @@ export default async function handler(
       .status(200)
       .json({ status: true, statusCode: 200, message: "success", data });
   } else if (req.method === "PUT") {
-    const { id, data } = req.body;
-    await updateData("users", id, data, (result: boolean) => {
-      if (result) {
-        res
-          .status(200)
-          .json({ status: true, statusCode: 200, message: "success" });
-      } else {
-        res
-          .status(400)
-          .json({ status: false, statusCode: 400, message: "failed" });
-      }
-    });
-  } else if (req.method === "DELETE") {
     const { user }: any = req.query;
-    const token = req.headers.authorization?.split("")[1] || "";
-    console.log(decoded);
+    const { data } = req.body;
+    const token = req.headers.authorization?.split(" ")[1] || "";
     jwt.verify(
       token,
       process.env.NEXTAUTH_SECRET || "",
       async (err: any, decoded: any) => {
-        if (decoded) {
+        if (decoded && decoded.role === "admin") {
+          await updateData("users", user[1], data, (result: boolean) => {
+            if (result) {
+              res
+                .status(200)
+                .json({ status: true, statusCode: 200, message: "success" });
+            } else {
+              res
+                .status(400)
+                .json({ status: false, statusCode: 400, message: "failed" });
+            }
+          });
+        } else {
+          res
+            .status(403)
+            .json({ status: false, statusCode: 403, message: "access denied" });
+        }
+      }
+    );
+  } else if (req.method === "DELETE") {
+    const { user }: any = req.query;
+    const token = req.headers.authorization?.split(" ")[1] || "";
+    jwt.verify(
+      token,
+      process.env.NEXTAUTH_SECRET || "",
+      async (err: any, decoded: any) => {
+        if (decoded && decoded.role === "admin") {
           await deleteData("users", user[1], (result: boolean) => {
             if (result) {
               res
@@ -51,7 +64,7 @@ export default async function handler(
         } else {
           res
             .status(403)
-            .json({ status: false, statusCode: 403, message: "access Denied" });
+            .json({ status: false, statusCode: 403, message: "access denied" });
         }
       }
     );
